# packages/api-client/src/model/song.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Music Collection API
 * The Music Collection API description
 * OpenAPI spec version: 1.0
 */

export interface Song {
  /** The artist of the song */
  artist: string;
  /** The URL/path to the song audio file */
  audioUrl?: string;
  /** The unique identifier of the song */
  id: string;
  /** The URL/path to the song image */
  imageUrl: string;
  /** The name of the song */
  name: string;
}